Import standalone toolchain doc from DAC.
Rather than trying to convert all the docs at once, let's just
migrate them one at a time. I'm starting with the standalone
toolchain page since that is the one that needs the most work.

This is pretty much an as-is translation from HTML to Markdown.

There's a makefile in docs/ that, when run, will invoke pandoc
(available from apt) to turn all of the markdown user documentation
into HTML to be imported into DAC. This way we can publish current
documentation (for a beta release, for example) to the NDK GitHub
page while still able to push the same documentation to DAC when the
stable release is available.

I've also added `scripts/update_dac.py` that will handle building and
copying the new ones into DAC. It will not remove the old
documentation, nor will it update the `_book.yaml` (navigation). This
is done so we can migrate on doc at a time and still leave the old
ones in place.

Bug: http://b/27166093
Change-Id: Ia090b7c3eb252d8af3e7cba322ed68f9325224cc