Stop building multilib gnustl.
We don't support any of the multilib configurations (x32, mips32r2,
mips32r6), so don't waste time building them.

    # Before this change:
    $ time ./checkbuild.py --module gnustl
    real    22m24.606s
    user    32m50.704s
    sys     14m5.013s
    $ du -sh ../out/dist/android-ndk-r12-beta1-linux-x86_64.tar.bz2
    660M    ../out/dist/android-ndk-r12-beta1-linux-x86_64.tar.bz2

    # After this change:
    $ time ./checkbuild.py --module gnustl
    real    13m23.073s
    user    19m34.126s
    sys     8m36.052s$ du -sh
    $ du -sh ../out/dist/android-ndk-r12-beta1-linux-x86_64.tar.bz2
    605M    ../out/dist/android-ndk-r12-beta1-linux-x86_64.tar.bz2

Note that some of the time and space we're getting back is because
we're now only building the things we do support once. Previously
building x86_64 also build x86, which had already been built
separately. The same was true for mips64 (though it was an duplicate 3
ABIs instead of one).

Bug: http://b/23437346
Bug: http://b/28214235
Change-Id: I2a552f3ceab63db89b25a6a13ac9caf7ffbf09fb