Fix a race in the parallel build.
We run `launch_build` from multiple processes. There can be a race
between the check for the log directory and the creation of it.
`os.makedirs` fails if the directory already exists, so that causes
it to fail.

Move the creation up to the parent process to avoid the issue.

Change-Id: Iac5f120c300db66c440a34b403af458f60bddce4