Stop compressing our intermediate packages.
The zips that pass through these two functions are only a part of the
modular NDK vaporware. We're compressing the hell out of them even
though they are never used. Just store them uncompressed to save time
in the build.

Before this change:

    Build: 0:14:16
    Packaging: 0:04:41
    Testing: 0:15:44
    Total: 0:34:43

After this change:

    Build: 0:13:38
    Packaging: 0:04:04
    Testing: 0:15:47
    Total: 0:33:31

Saves about a minute (the build step is affected by this too since
that's the part that creates the zips).

Bug: http://b/28214235
Change-Id: I7027aaa38efc78a0bdc4d2418e0f55f03360010b