Fix make-standalone-toolchain's package mode.
If `make-standalone-toolchain.sh` is invoked without an
`--install-dir` argument, it is supposed to create a tarball from the
toolchain.

https://android-review.googlesource.com/#/c/202472/ changed the name
of the directory we build the toolchain from to be a temp dir so we
don't collide with simultaneous builds. Since the directory name is
now random, we can't build the tarball properly.

Build everything in a subdir of the temp dir with the same name as
the toolchain.

Bug: https://github.com/android-ndk/ndk/issues/88
Change-Id: I7911f07a7c34c0934194546dafe184ed8fe9b07b