Specify toolchain library path when using clang for mips32
Clang does not yet support mixing mips32 target and mips64el gcc. Until
this is fixed, add -L flag with path to gcc's lib folder, to allow build
to pass.

Issue17144-byteswap test is specific as it tries to target mips32r2.
Instead of fixing it in the same way, just disable the test for now.

Once clang is fixed this change can be reverted.

Change-Id: Id2d5c3374103a2857d4b118b93cb5097e418bba3