Convert make-standalone-toolchain to Python.
Aside from a much needed cleanup, this lets our Windows users use
standalone toolchains without needing to install cygwin.

This creates identical toolchains to the old tool, which is why it
has so much crap. I'm not sure if we want to release it as is so it's
a drop in replacement, or release it cleaned up. I'm tempted to do
the latter and ship both versions for r12, and we'll remove the old on
in r13.

Change-Id: I6b98557b4dbff161df692ab95e1c4bd81f1616d1