Fix make-standalone-toolchain.sh for libc++.
We don't have both ARM and thumb libs any more, just thumb. This will
install the thumb libs to the main directory, but I think that's what
we want long term anyway (will have to do some testing with gcc to
make sure that's right, but this at least gets it running again).

Change-Id: I4f1e0f4cd003531b7a38531f7d13c2ebb534a82b