Fix libdir for STL libs.
What we had here before was actually identical to
`make-standalone-toolchain.sh`, even in r10e, yet in r10e it was not
a bug. Why? Because of never ending hacked up Clang landmines:
https://android.googlesource.com/toolchain/clang/+/f3846c4cc9d3d238e35b2a4ac506d2a047ee44c5

Note that the above patch was a deviance from GCC behavior. The
current Clang is the correct one.

Change-Id: I4c53c5028c79700ed715b05ae31bce5d95e98418