Replace build tuples with a Module class.
This paves the way to cleaning up our builds. Right now we don't have
a good way of keeping metadata for builds (such as install
locations or dependencies), and a lot of this information gets
duplicated and split into too many places.

Right now it's just the minimal migration to a new format that uses a
Module class rather than an arbitrary build function, but more will
be added in later patches.

Test: ./checkbuild.py
Bug: None
Change-Id: I8b89c23317ce64bd78e6e3199793d95392aedbc9