Clean up API level in test_config.py.
The API level was being passed as the string "android-$API_LEVEL",
which meant you had to perform some pre-processing before using it.

Pass it in as an integer instead. To avoid accidentally comparing it
with `None` (integer comparisons with `None` are generally not a good
idea in python2, and invalid in python3), stop allowing `None` for
this value. Check Application.mk for a default, or fallback to the
default level for the ABI.

Change-Id: Ib97f544d0e4b4d4d0798b5d82e749fc8caaa8a9f