ndk-test: fix toolchain path for x86 and x86_64
without this patch clang does not find the correct ld and starts using
/usr/bin/ld. Furthermore clang will fail to find the correct -lgcc.
The problem is that the lit config points to a dir name that does not exist.
The name of the toolchains installed when building with checkbuild.py are:

$ ls out/android-ndk-r13-beta1/toolchains
aarch64-linux-android-4.9 arm-linux-androideabi-4.9 llvm
mips64el-linux-android-4.9 mipsel-linux-android-4.9 x86-4.9 x86_64-4.9

The patch corrects the toolchain path names for x86 and x86_64.

Change-Id: I679a6b88628e89c9c090af7e6be54e6b391494b4