Add LOCAL_ASFLAGS and APP_ASFLAGS.
We had LOCAL_ASMFLAGS and APP_ASMFLAGS, but those only applied to
.asm (yasm) files. There was no way to pass options to .S files but
not .c files.

Add LOCAL_ASFLAGS for module specific .S only flags, and APP_ASFLAGS
for application wide settings.

Note also that APP_ASMFLAGS (the yasm one), though clearly someone
tried to implement it, didn't actually work because ASMFLAGS was
never added to the list of app specific variables. The test I've
added caught this, so the fix is included here.

Test: Added build/asmflags for verifying that .c, .S, and .asm files
      all get the right set of flags attached.
Bug: https://github.com/android-ndk/ndk/issues/150

Change-Id: If495edf4e589468a6e46ef01553641d97e18093a