Install support headers to usr/local/include.
Historically these were installed to the same directory as the C++
headers, but with the updated libc++ we have copies of a lot of those
headers in libc++ itself that we end up clobbering.

This problem should go away with unified headers, but those aren't
ready yet. For the time being, install the libandroid_support headers
to a different builtin include path. usr/local/include seems to be the
least objectionable option.

Test: Added `#include <cmath>` to the standalone toolchain test;
      mathias's project builds again :)
Change-Id: Icd036617c43cc6cbb18a776e8927f007afd5ee22