Fix using libandroid_support with c++_shared.
First problem: we weren't linking libandroid_support when using
libc++_shared.

The second problem is a bit harder to solve: we need to get the
include ordering between libc++ and libandroid_support correct to
make sure the correct headers are overridden. We do this by not
exporting the libandroid_support includes.

This is a bit of a hack, but we need to make sure these come *after*
libc++, and we don't have a way to tell ndk-build whether the libc++
headers (which include the libandroid_support headers for this
reason) take priority over the libandroid_support headers.

Since we don't really support using libandroid_support independently
of libc++, just stop exporting the headers here and let libc++ handle
it.

Test: Added tests/build/log2.
Bug: http://b.android.com/212634
Change-Id: Id8853a07564a9e3486c520ccf6c0e1f2c4ab9a89