Always build the vulkan test.
`match_unsupported` is invoked multiple times for device tests. Once
with `subtest=None` to check if we even support building this project,
and then once for each executable within the test to see if we support
running the test for this configuration.

We can always build this test regardless of what device we have
attached, so gate the check on `subtest != None`.

Change-Id: Icf08aa148ea8a602c6b7dc14f9e8c500831e0378