Fix `ndk-build clean` for executables.
I'm not really sure if this ever worked before. There's some stuff in
build/core/build-all.mk that looks like it was supposed to handle
this, but it's all bogus:
https://android.googlesource.com/platform/ndk/+/master/build/core/build-all.mk#144

`EXECUTABLES` and friends are never defined, so that does nothing.

This is a pretty minimal fix. This leaves the empty obj and libs
directories. If I could go back in time I'd just make the clean target
blow these away, but I worry about changing that at this stage in
case people are actually keeping things in there...

Test: ndk-build && find libs && ndk-build clean V=1 && find libs
Bug: https://github.com/android-ndk/ndk/issues/80
Change-Id: I1747289439612da82b5a9611076341d3052b7e8e