Print a test failure report after the build.
Previously we printed the test failures at the end of each ABI, but
when performing a full build this made it difficult to find failure
details in the full log unless they happened for x86_64 (the last ABI
to be tested).

After running all tests, group all test failures by name and print a
report. This duplicates the result printing in the case of only
building a single ABI, but that's better than not printing enough. I
probably need to take a pass through here and clean up all the test
reporting at some point.

Also write this report out to build_error.log so we get a concise
description of why the build failed. We've been doing this for build
failures, but test failures required scanning the entire multi-MB
build.log (which takes several minutes to load in chrome on a mac).

Test: Added #error to a test, ./checkbuild.py.
Bug: None
Change-Id: Ie08f9d86e82eeaac544b578fcf8a31087c3c50e1