Clean up the build parallelism.
While multiprocessing.Pool is nice for things like map, it really
sucks for anything that needs to spawn it's own subprocesses.
SIGINTing the build would result in a cascade of stack traces from
all the child processes (and nothing seemed to be able to silence
those). It also did not always die on the first SIGINT.

Replace the whole thing with a WorkQueue based on more primitive
multiprocessing types.

Test: nose2
      ./checkbuild.py # interrupted
      ./checkbuild.py # run to completion
Bug: None
Change-Id: I08a08b371e07b49269963085c9ed93d4f2e7b442