Put scripts in the right place.
Imports in python are relative to the script that was invoked. We
want to be able to share things between build and test without having
to resort to hacks like site.addsitedir, so put the tests scripts in
the root.

I've left tests/run-all.py as a proxy script for the real
implementation which now live in the root as run_tests.py. I didn't
add proxy scripts for validate.py or test_libcxx.py since I'm the only
one that runs those anyway.

Test: ./validate.py path/to/ndk
      ./run_tests.py path/to/ndk
      ./tests/run-all.py path/to/ndk
      ./test_libcxx.py path/to/ndk
Bug: None

Change-Id: I03e562ea8bcfc3d3191db76505f9ba6b4f208918