Make Test pickleable.
Right now I can't use WorkQueue for parallelizing the tests because we
instantiate all the tests before running them, and we can't pass a
Test through a multiprocessing.Queue because it isn't pickleable.

The config has properties that are initialized to either the module's
definition of the config method or one from NullTestConfig. Methods as
values can't be pickled.

Stop caching the test config module. Python caches the import for us
anyway, so there's no need to do this.

Test: ./validate.py
Bug: None
Change-Id: I7ba13e56367658befef37ad6bcc99ffceefaba17