Parallelize the test runner.
For ./run_tests.py --abi armeabi-v7a this cuts the run time from
3m40s to 50s. For ./validate.py its from 70m to 15m.

There's still more that can be improved here. If the test suite was
just one project with all the tests then we'd have no parallelism
since we're only assigning jobs per project. For the case of
validate.py, we're still running all the configurations in serial.

Test: ./run_tests.py --abi armeabi-v7a && ./validate.py
Bug: None
Change-Id: I6d5cf9ee3e1b0c087b9503bea02f69d4140436ff