Re-raise worker exceptions in the calling process.
Uncaught exceptions in a worker process will be caught and packaged
for transport to the caller, and then re-raised if retrieved.

Unfortunately tracebacks are not pickleable, so we can't re-raise the
exact exception we caught. We have to settle for raising our own
exception whose message is the traceback from the original exception.
We'll see a traceback for where the exception was re-thrown from as
well as the trace for the original exception, but the type of the
re-thrown exception will always be TaskError.

Test: nose2
Bug: None
Change-Id: Id0f3933e913e3f8b37c88c2c1567f217305ff564