Flush queues so a terminated WorkQueue can join.
If there are still items pending in the queues when terminate is
called, the subsequent join will hang waiting for the queues to be
emptied.

I haven't been able to come up with a test that exhibits the failing
behavior. I've only seen this when running the tests and an exception
was raised in the worker.

Test: nose2
Bug: None
Change-Id: I35fb943023d734c61d70ccee1abe4f5a7e385345