Fix race in header build tests.
This test wasn't honoring APP_ABI beyond the architecture, and was
building to a well known location based on the platform and ABI. If
this was run for both armeabi and armeabi-v7a at the same time,
they'd both try building each of those architectures and collide.

Stop using a well known directory under TMPDIR and start using the
build directory that the test runner sets up for us.

Also, as a speedup (this is one of the longest running tests), only
build for the ABI that we've actually chosen, not all ABIs in the
architecture for that ABI.

Test: ./run_tests.py --filter *-prctl --abi armeabi-v7a
Bug: None
Change-Id: Icf50b5ab27908efefdcb8a88ae10d217acd8a5d5