Fix libandroid_support's math_private.h for arm5.
libandroid_support's definition of the byte-order for doubles on
armeabi was wrongly big endian rather than following the byte order of
the ABI. This was due to a missing check for __ARM_EABI__ (a bug in
the upstream FreeBSD libm source that bionic did not have because it
has been updated since it was fixed).

I haven't pulled the full update of math_private.h since it also
needs a lot of new macros for complex math support. I'll be cleaning
up libandroid_support more thoroughly soon anyway.

Note that this bug could surface even for arm7 when using a
standalone toolchain because we still have the Clang bug where it
doesn't search architecture specific lib directories, meaning that
arm5 gets used instead of arm7 (http://b/26180518 and
https://github.com/android-ndk/ndk/issues/151).

Test: ./run_tests.py --filter log2
Bug: https://github.com/android-ndk/ndk/issues/204
Change-Id: I2caaeac3afe0830008b865bc124f48200ac5f4b6