Fail if a module has missing dependencies.
    include $(CLEAR_VARS)
    LOCAL_MODULE := foo
    LOCAL_SRC_FILES := foo.cpp
    LOCAL_STATIC_LIBRARIES := never-defined
    include $(BUILD_SHARED_LIBRARY)

In the above example, the build would succeed even though it had
missing dependencies. This seems to have always been the behavior (I
checked back to r10e), so there's a good chance that there are builds
out there that depend on this behavior.

By default we will now abort in this situation because this is so
completely broken. A user may define `APP_ALLOW_MISSING_DEPS` to
"false" in their Application.mk or on the command line to revert to
the old, broken behavior.

Test: ./run_tests.py --filter APP_ALLOW_MISSING_DEPS
Bug: https://github.com/android-ndk/ndk/issues/208
Change-Id: Ie77ad18a7e8000ef468c79f8e89aa2e5ae1788dd