Add support for is_negative_test.
Note that this is different from build_broken. Use build_broken to
indicate a bug and use is_negative_test to indicate a test that should
fail if things are working.

Also note that check_broken and is_negative_test can be layered. If a
build is expected to fail, but doesn't for armeabi, the test_config
could contain:

    def is_negative_test():
        return True

    def build_broken(abi, api, toolchain):
        if abi == 'armeabi':
            return abi, bug_url
        return None, None

Test: Added a test that uses this in the next patch.
Bug: None
Change-Id: Ib12e974db9bc58d9e28ec97646cfb81c208cc4cd