Fix some paths in ndk-build.
These don't need to be correct for normal operation of the NDK, which
is why we haven't had issues about these yet.

The first check is for Cygwin using a 32-bit NDK on a 64-bit system (I
think).

The second is to use the prebuilt make instead of the system's. Since
any Linux/Darwin developer machine is already going to have make
installed, we hit the fallback case. Using mingw requires a mingw
make anyway, and Windows doesn't use this script.

Test: ndk-build V=1 # check that we're using the prebuilt make
Bug: None
Change-Id: Ia7dc9af1031bdeb6788b0b8c2ea0b33872a9f80e