Add missing _GNU_SOURCE define in math test.
It's C, so the compiler is perfectly happy to let the user call an
undefined function. On top of that, the old NDK headers have
_GNU_SOURCE defined unconditionally in features.h.

The really fun part: without this change, this test segfaults for
armeabi-v7a when compiled as C when using unified headers. Something
seems to be wrong with the builtin definition in Clang?

Test: ./run_tests.py --force-unified-headers --filter math
Bug: None
Change-Id: I0013efaae76c1f050f0ca8c410b68ebebcfd1213