Topologically support imported properties.
We need to guarantee that libc++ includes are included in the right
order with respect to libandroid_support includes (libc++ ->
libandroid_support -> libc). In general, a dependent module's headers
should appear first on the command line so they can override those
supplied by the dependency.

Before this patch, we did this by exporting the libandroid_support
includes from libc++ rather than from libandroid_support itself. This
has the unfortunate side effect of making it impossible to get the
libandroid_support includes if using *only* libandroid_support (as a
sort of proto-compatibility library). We need this to work since
we're relying on libandroid_support to provide fenv functionality for
unified headers.

Test: ./run_tests.py --filter log2
Bug: http://b.android.com/212634
Change-Id: Icaf7e5d54bb3369484744f7f7c7c9aedcbb6a09d