Support unified headers in standalone toolchains.
These aren't supported as smoothly for GCC. Unified headers require
that `__ANDROID_API__` be defined from the command line. For Clang we
do this in the wrapper scripts that are used to pass the target, but
we don't have wrapper scripts for GCC. Since GCC is deprecated, just
don't bother.

Test: ./run_tests.py --filter standalone-toolchain*
Bug: https://github.com/android-ndk/ndk/issues/120
Change-Id: I71d42d218b83bbe8634a5e473804c9e6d7199101