Move our earliest testing target to android-15.
With gmscore dropping support for Gingerbread, we can now start using
ICS as a minimum target. This doesn't drop support for Gingerbread
entirely, but that will follow soon.

Note: gmscore actually moved the minimum target to 14, not 15, but
https://developer.android.com/about/dashboards/index.html shows that
there are only a miniscule number of android-14 devices out there,
and there are no Nexus device images for 14 available, so there's
nothing to test on anyway.

The libc++ tests finish on a Galaxy Nexus in 7m30s, compared to never
having successfully finished on Nexus One.

Test: ./validate.py
Bug: None
Change-Id: Id03139925e6a128f8bcc85520941848516252bd1