Always install a module if it is built.
With this, rebuilds of single components are much faster. Rather than
needing to `./checkbuild.py --module clang --force-package` and then
wait for all modules to be reinstalled and the package recreated,
`./checkbuild.py --module clang` will now replace the clang module in
the existing install.

NB: If a module's install path is changed, the old module will *not*
be removed. Remove it manually or do a clean build. This is only a
problem for local builds; the build servers always perform clean
builds.

Test: touch build/bar && ./checkbuild.py --module ndk-build && \
          ls ../out/android-ndk-r15-canary/build/bar
Bug: None

Change-Id: I2175d3bc1c2afb9dacba44d23aeec21d74b856f2