Add a warning about our languishing ndk_helper.
Turns out this isn't usable without a Java component we don't have
anyway. Just accept that this one is essentially dead and point
people at the maintained one instead.

We'll leave this copy around for compatibility for now, but we'll
probably try removing this from a beta 1 in the future to see if
anyone is actually using it.

Test: Made a sample project that uses this, ran ndk-build, saw the
      warning. No warnings if not used.
Bug: None

Change-Id: Iae62eb5f5a2641bf5eace3ad247f1df7db4594c4