Push the whole test directory, not one at a time.
adb pushing each file individually is a lot slower than pushing the
whole test directory. The downside is that we have to push the whole
directory even if a test filter means that we only need one
executable (we can fix this in a follow up by looking at the test
filter).

This saves us a couple minutes in a validate.py run (22 minutes down
from 25).

Test: ./validate.py
Bug: None
Change-Id: I8b7e34106e84389cb63ff260af7019d98f4de277