Remove gen-cywgin-deps-converter.awk.
As far as I can tell we don't need this at all. For one, Cygwin users
can invoke ndk-build.cmd (and in fact, at the moment that's their
only choice out of the box) and in that situation HOST_OS is
"windows" rather than "cygwin" and this code doesn't get run at all.
If you manually create the ndk-build (shell) shortcut for cygwin,
this isn't needed anyway since all the paths in dep files seem to be
relative paths.

At some point we should figure out if we need the cygwin support at
all. As far as I can tell it's just a bunch of hacks that we don't
need.

Test: ./run_tests.py --abi armeabi-v7a  # In cygwin
Bug: None
Change-Id: Iff080a0bfdf318f781d0a4ca6455b36c0dbaea06