Mark stof test as xfail.
The C++ spec says that std::stof throws std::out_of_range if strto*
sets errno to ERANGE or if the value is beyond the representable
range. The C spec says that strto* setting ERANGE is implementation
defined. Apparently glibc does and we don't, and libc++ is relying on
the implementation defined behavior here.

Test: ./run_tests.py --suite libc++ --abi armeabi-v7a
Bug: http://b/34739876
Change-Id: If62b7c3c072bf658782eeb85dbec594001d50516