.note.android.ident should be type NOTE, not PROGBITS.
The note changed from .note.ABI-tag to .note.android.ident in 2012,
but the script was never changed to match.

The comment in crtbrand.c seems to be wrong too; bionic has a crtbrand.S
that works just fine. We should come back and remove this nonsense and
just build direct from a .S file. But this fixes the immediate bug for now.

Bug: http://b/34779141
Test: added test to gen-platforms.sh itself; still builds
Change-Id: Id43d21fcf2fd3c79f5b867b9c48005e944eb3bbd