Make unified headers the default.
With unified headers shipped in r14, it's time to make them the
default in r15. Flip the defaults for the various build systems and
change the language  for the docs and options to make it more clear
that the alternative is deprecated.

This also changes the name from APP_UNIFIED_HEADERS to
APP_DEPRECATED_HEADERS. For people that opted in to unified headers,
this won't make any change since they're now on by default. For
people that have already opted out, this will serve as a reminder
that they need to look in to this.

Test: ./validate.py
Bug: None
Change-Id: I8af74cf72f9c50ecab01b0d61f9e38a2c9a1d575