Lock the headers include test to the old headers.
I need to rewrite this test to do anything useful with unified
headers, so just lock it to the old ones for the moment.

I've also removed the completely unused jni directory in the test (it
took me way too long to figure out why my Application.mk changes
weren't doing anything), and renamed the test to actually say what
it's testing.

Finally, I stopped sending the output to /dev/null so we can actually
see what goes wrong when it does.

Test: ./run_tests.py --abi armeabi-v7a --filter headers-include
Bug: None
Change-Id: Idc3458741c499e53dd37f2da8b0cc028b5b2c86c