Mark nan strto* tests as xfail for lp64.
The platform didn't actually get nan support in strto* until very
recently. We'll probably fix this with libandroid_support (that's
what we have for lp32), but I don't want to add more of the current
libandroid_support to do that, so I'm marking it as xfail until I can
replace it with bionic's source.

Test: ./run_tests.py --suite libc++ --abi arm64-v8a
Test: ./run_tests.py --suite libc++ --abi armeabi-v7a
Bug: http://b/31101647
Change-Id: Id27b41ea356d6c405ab2d5ec5e25baaf25896ede