Fix notice checking for PackageModule.
This wasn't being run for PackageModule types. Nothing regressed
aside from missing repo.prop files, but all the ones missing one
would match the top level one anyway.

`create_repo_prop` is a new property for generating a repo.prop file
for the current build. This defaults to off since we want to ship the
one that matches the prebuilt if there is one.

Test: ./checkbuild.py
Bug: None
Change-Id: I396a9b37ece77299b50c38da929ee3d20f392799