Fix libc++ test filtering.
We need to have the exact paths here for filters to work properly. If
we have a filter like

    --filter libc++.std/strings/string.conversions/stold.pass.cpp

we'll run that test, but then we'll skip "running" the XunitResult
"test" because it won't have the same name; the "string.conversions"
will be "string_conversions" and not match the filter.

Because of this, rename libc++ test configs to match real paths.

Test: ./validate.py
Bug: None
Change-Id: Ib478747c78ec3b1c642f36c3e385f045a9277d7f