Appease nose2's test discovery code.
All our test configs are unfortunately named test_config.py, which
matches the test_*.py glob that is expected to be a python test.
nose2 finds multiple modules with the fully qualified name of
test_config and explodes. Not sure why we don't need to add these for
every test_config.py directory, but it seems we don't.

Test: nose2
Bug: None
Change-Id: Ia4465cba0b6cf3d47f5f75e9722e1e22da1092d9