De-dup test names.
There's actually a race here, we've just never managed to hit it. One
of these tests (fenv) was actually a duplicate, while the other was
not well named. Fix that.

Another alternative for fixing the race is to include the
subdirectory in the test build directory name, but that information
isn't plumbed very deep, and enforcing that there are no duplicate
test names makes `--filter` easier to use anyway.

I've added enforcement of no duplicate build directories, but that is
coming in a later patch.

Test: ./validate.py --filter fenv
Test: ./validate.py --filter log2
Test: ./validate.py --filter log2-libcxx
Bug: None
Change-Id: Ifeaa3fd224ec4dcb714137b909f64ae9398525f4
(cherry picked from commit e324a9a8057d170da63d81d9e95485839c79b8d6)