Get Darwin Clang target libs from Linux.
We don't build target binaries as part of the Darwin build. The Darwin
toolchains need to get these from the Linux prebuilts.

The headers and libraries we care about are all in lib64/clang for
both toolchains, and those two are intended to be identical between
each host, so we can just replace Darwin's with the one from the Linux
toolchain.

Bug: http://b/37915581
Test: ./run_tests.py --filter test-openmp --abi armeabi-v7a
Change-Id: Ibb47e40ddb30a3eacef9d715fee7e2633557980d