Hide libgcc in all binaries.
It seems we can't guarantee the link order of our libraries (objects,
then static libraries, then shared libraries) in cmake, which
prevents us from ensuring that libunwind is used with libc++ in the
event that some other library is exporting an unwinder.

The best we can do without patching cmake (and until the new unwinder
is available) is hide libgcc in any new libraries we build. This
can't protect us from libraries built without this patch (or with
build systems we don't control), but it will at least protect us from
ourselves.

Patch this in both ndk-build and cmake. ndk-build doesn't have the
same problems that cmake does, but libraries built by ndk-build can
cause problems for cmake. Standalone toolchains are also affected by
this. I'll fix them in a follow up patch.

Test: Built the test cast attached to the bug, libgcc is hidden.
Bug: https://github.com/android-ndk/ndk/issues/379
Change-Id: I3406c7e49f7194231bfe1e5f921a3e51d875dc84