Get rid of the test logs directory for validate.py.
The test logs aren't particularly useful as-is. These were necessary
back when we didn't have good test summary printing, but that was
fixed ages ago. These are not just overly verbose versions of what we
already have, and it's not obvious which logs actually contain
failures anyway.

Just use a less verbose StdoutPrinter instead of logging to a file.
Only errors will be printed to stdout.

Keep the argument for the moment since CI expects it to exist.

Test: ./validate.py
Bug: None
Change-Id: I3986a8e19698a58795492dd0f6e8bfb2fb75e92d