Fix ndk-gdb for the Samsung Galaxy S8.
Third-party developers were reporting that they couldn't debug on the
Samsung Galaxy S8. They assumed it was because of a run-as problem like
the one we'd seen on earlier Galaxy devices, but it's not. run-as works
fine on the S8. The problem is actually that the device seems to ship
with a property "ro.csc.facebook.partnerid" whose value contains a
newline. I assume the newline is a bug (especially coming at the end),
but there's nothing we can do about it now (and newlines are allowed in
property values).

Move ndk-gdb.py off `get_props` which used `getprop` and then tried to
parse all the properties into a map, and use `getprop X` for any property X
that we actually need. It's not noticeably slower (on a modern device like
the S8, at least), and it's more robust.

Bug: https://issuetracker.google.com/37093233
Test: ndk-gdb on a Samsung Galaxy S8
Change-Id: I8b289578c467bab40008f27008679c253761e212