Fix test broken config for strto* tests.
Some of these have started passing for lp64 because libc++ now calls
strto* instead of sscanf, but the strtold still has a broken nan
parser until very recently.

For lp32, since we're now calling strtof, we end up calling strtod
because strtof was an inline function that deferred to strtod. The
number we try to convert in the test *is* representable as a double,
so the conversion doesn't fail.

Test: ./run_tests.py
Bug: http://b/31101647
Bug: https://github.com/android-ndk/ndk/issues/415
Change-Id: I02c2c23815548052850c0b2f3132f335c6cab086