Rewrite gen-platforms.sh in Python.
The shell script was more difficult to make changes to, and since the
new implementation doesn't rely on copying each platform as it moves
forward, nor zipping the archive just to extract it a moment later,
this is slightly faster.

I haven't update build/tools/README.md because frankly that doc needs
to be deleted, but I need to take a look through it to see if there's
anything worth preserving first.

Test: diffed new directory against old directory (layout only)
Test: ./run_tests.py
Bug: None
Change-Id: Iad04b3ca49334462e26178757dcf28a21ae4cbae