Add _FILE_OFFSET_BITS=64 support for libc++.
Including things like cstdio requires that we have a decl for
everything that header needs to pull into the std namespace. We don't
actually have these functions available in this release, but we can
at least make this header includable.

Adding a test to make sure all of the C++ C headers can be included
in _FILE_OFFSET_BITS=64 mode.

Test: ./run_tests.py
Bug: https://github.com/android-ndk/ndk/issues/480
Change-Id: Icfd319b293736249e5840fee7efbce1d3b69e20b