Allow sharding tests across identical devices.
Might be worth loosening our definition of what's allowed in a group,
but I'll start with the conservative match.

This could be improved a bit more. Tests are assigned to devices up
front here. With a little more work we could assign them to devices
as they finish their previous run so every device remains active as
long as possible in the event that one shard segment has more long
running tests than the others.

Test: ./run_tests.py -v
Bug: None
Change-Id: I8a9258903c323ba1c7550e101127dd0c2b3537da