Use multiprocessing.Manager.Queue.
This seems to get rid of most of the issues surrounding interrupting
worker processes. multiprocessing.Queue actually spawns a thread that
pushes the data across the pipe, and interrupting that thread at the
wrong time seems to put it into a bad state where it locks waiting to
receive more data that it will never get.

multiprocessing.Manager.Queue doesn't have these constraints (at
least according to the docs).

Test: ./run_tests.py  # SIGINT, repeat
Bug: None
Change-Id: Ibda7e6a0304cd870f61605237462c0b305071dc6