Configure macOS ndk-stack for arm64 rather than arm32
I verified that arm32 "objdump -x" and arm64 "objdump -x" output the same
thing for libc.a across various archs. I think libbfd is configured for
a particular triple and isn't generally expected to work with binaries of
a different architecture, so this fix isn't ideal, but it's also trivial.

Bug: https://github.com/android-ndk/ndk/issues/20
Test: run ndk-stack manually on traces of various archs
Change-Id: I2c4783ef2f7fec2af0f1a7671c62a47309f7e819