Fix the PIC code in the x86 yasm test and add x86_64 support
This instruction was segfaulting:

    mov  [esp], dword fmt wrt ..gotoff

The build-time linker replaces "fmt wrt ..gotoff" with an offset from the
end of the GOT to the fmt symbol. We need to setup a PIC register (in ebx)
containing the end address of the GOT, then add ebx to the fmt gotoffset.

Also: extend the test to x86_64 to fix the buganizer bug.

Bug: http://b/24620865
Test: ./run_tests.py --filter yasm --abi x86 --abi x86_64 --rebuild
Change-Id: I60dd0f594bf6b5a9b48385808687513653035f2f