Hide android_support's UniquePtr class
Currently, the reset function is generated out-of-line in
libandroid_support.a. An app can declare its own unrelated UniquePtr
class, and if it happened to declare a reset method with the same
signature, it would override the one used by android_support's
wcstox.cpp and swprintf.cpp files.

The problem symbols only appear for the armeabi-v7a architecture (and
armeabi in r16-beta2).

e.g. We currently have these two weak symbols:

$ nm -S armeabi-v7a/libandroid_support.a | c++filt | grep UniquePtr
00000001 00000018 W UniquePtr<char [], DefaultDelete<char []> >::reset(char*)
00000001 00000018 W UniquePtr<char [], DefaultDelete<char []> >::reset(char*)

After this change, we have these symbols instead:

$ nm -S armeabi-v7a/libandroid_support.a | c++filt | grep UniquePtr
00000001 00000016 t (anonymous namespace)::UniquePtr<char [], (anonymous namespace)::DefaultDelete<char []> >::~UniquePtr()
00000001 00000016 t (anonymous namespace)::UniquePtr<char [], (anonymous namespace)::DefaultDelete<char []> >::~UniquePtr()

Test: ./checkbuild.py && ./run_tests.py
Change-Id: Ic7d34e910f83adde687351f5f83d11dd49024e0a